refactor: please read detail
- add new package lotto to wrap the application
- separate logic in controller
- adapt LottoException in retryUntilSuccess
- documented domain classes
