--- src/main/kotlin/Application.kt
@@ -1,6 +1,0 @@
-import controller.LottoController
-
-fun main() {
-    val controller = LottoController()
-    controller.run()
-}
