--- src/main/kotlin/util/LottoUtils.kt
@@ -1,15 +1,0 @@
-package util
-
-object LottoUtils {
-    fun <T> retry(
-        block: () -> T,
-    ): T {
-        while (true) {
-            try {
-                return block()
-            } catch (e: IllegalArgumentException) {
-                println(e.message)
-            }
-        }
-    }
-}
